Change Page Structure To Be Side By Side
Change the page structure to be side by side because most important
information will be on the right giving the user details of what is in
the resume. While the basic information will be on the left to allow the
user to contact me at any point, on the page, as it scrolls with them.