Add Content & Format To Right Side, Fix CSS Names
Add in the content and format them tot he right side, of the page,
because the content needs to be presentable, to the user.

Fix the name to be more manageable and allow the page to load a little
bit faster because id actually slow down the site even for a split
second and makes CSS not as cosing as it should be.